ag-associates-ai/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
langgraph==0.0.29
langchain==0.1.0
langchain-community==0.0.10
openai==1.7.2
httpx==0.26.0
numpy==1.26.3
